Revert "Merge branch 'VOREStation:master' into master"
This reverts commit 624dfd6188569f31e90de9bf7f78e81dc6cc368f, reversing
changes made to b880a326ce5e9ebda61c6c7582d92c6f6b30c474.

+++ code/modules/hydroponics/grown.dm
@@ -162,7 +162,8 @@
 			M.Stun(8)
 			M.Weaken(5)
 			seed.thrown_at(src,M)
-			qdel(src)
+			sleep(-1)
+			if(src) qdel(src)
 			return
 
 /obj/item/reagent_containers/food/snacks/grown/throw_impact(atom/hit_atom)
@@ -354,7 +355,7 @@
 		if(!reagents || reagents.total_volume <= 0)
 			return
 		reagents.remove_any(rand(1,3)) //Todo, make it actually remove the reagents the seed uses.
-		var/affected = pick(BP_R_HAND,BP_L_HAND)
+		var/affected = pick("r_hand","l_hand")
 		seed.do_thorns(H,src,affected)
 		seed.do_sting(H,src,affected)
